chore: refactor into a loader component and a form renderer
