+++ src/modules/callouts/pages/CreateCallout.vue
@@ -3,51 +3,22 @@
     <PageTitle :title="t('createCallout.title')" />
   </div>
   <div class="flex gap-8">
-    <div class="flex-0 basis-menu">
-      <Stepper v-model="selectedStepIndex" :steps="stepsInOrder" />
-    </div>
-    <div class="flex-1">
-      <AppHeading class="mb-5">{{ selectedStep.name }}</AppHeading>
-      <component
-        :is="selectedStep.component"
-        v-model:data="selectedStep.data"
-        v-model:validated="selectedStep.validated"
-      ></component>
-
-      <div class="flex mt-5">
-        <AppButton
-          variant="linkOutlined"
-          :disabled="selectedStepIndex === 0"
-          @click="selectedStepIndex--"
-          >{{ t('createCallout.actions.back') }}</AppButton
-        >
-        <AppButton
-          v-show="!isLastStep"
-          class="ml-2"
-          :disabled="!selectedStep.validated"
-          @click="selectedStepIndex++"
-          >{{ t('createCallout.actions.continue') }}</AppButton
-        >
-        <AppButton
-          v-show="isLastStep"
-          class="ml-2"
-          :disabled="!isAllValid"
-          @click="submitForm"
-          >{{
-            steps.dates.data.startNow
-              ? t('createCallout.actions.publish')
-              : t('createCallout.actions.schedule')
-          }}</AppButton
-        >
-      </div>
-    </div>
+    <CreateCalloutForm v-if="steps" v-model="steps" />
   </div>
 </template>
 
 <script lang="ts" setup>
 import { parseISO } from 'date-fns';
 import slugify from 'slugify';
-import { reactive, ref, computed, markRaw, watch, onBeforeMount } from 'vue';
+import {
+  reactive,
+  ref,
+  computed,
+  markRaw,
+  watch,
+  onBeforeMount,
+  Ref,
+} from 'vue';
 import { useI18n } from 'vue-i18n';
 import { useRouter, useRoute } from 'vue-router';
 import { createCallout, fetchCallout } from '../../../utils/api/callout';
@@ -56,9 +27,7 @@ import type {
   GetMoreCalloutData,
 } from '../../../utils/api/api.interface';
 import PageTitle from '../../../components/PageTitle.vue';
-import AppHeading from '../../../components/AppHeading.vue';
-import AppButton from '../../../components/forms/AppButton.vue';
-import Stepper from '../components/Stepper.vue';
+
 import StepVisibility from '../components/steps/Visibility.vue';
 import StepTitleAndImage from '../components/steps/TitleAndImage.vue';
 import StepEndMessage from '../components/steps/EndMessage.vue';
@@ -67,11 +36,14 @@ import StepMailchimpSync from '../components/steps/MailchimpSync.vue';
 import StepDatesAndDuration from '../components/steps/DatesAndDuration.vue';
 import StepContent from '../components/steps/Content.vue';
 import { Steps } from '../create-callout.interface';
+import CreateCalloutForm from '../components/CreateCalloutForm.vue';
 
 const { t } = useI18n();
 
 const router = useRouter();
 const route = useRoute();
+
+const steps: Ref<null | Steps> = ref(null);
 
 const makeCalloutData = (steps: Steps): CreateCalloutData => ({
   slug: steps.url.data.useCustomSlug
@@ -199,52 +171,16 @@ const makeStepsData = (data: GetMoreCalloutData): Steps => ({
   },
 });
 
-//@ts-expect-error
-const steps: Steps = reactive(makeStepsData({}));
-
-// TODO: should just be a computed
-watch(
-  () => steps.titleAndImage.data.title,
-  (title) => {
-    steps.url.data.autoSlug = slugify(title, { lower: true });
-  }
-);
-
-const stepsInOrder = computed(() => [
-  steps.content,
-  steps.titleAndImage,
-  steps.visibility,
-  steps.endMessage,
-  steps.url,
-  //steps.mailchimp,
-  steps.dates,
-]);
-
-const selectedStepIndex = ref(0);
-const selectedStep = computed(
-  () => stepsInOrder.value[selectedStepIndex.value]
-);
-const isLastStep = computed(
-  () => selectedStepIndex.value === stepsInOrder.value.length - 1
-);
-const isAllValid = computed(() =>
-  stepsInOrder.value.every((step) => step.validated)
-);
-
-async function submitForm() {
-  const callout: CreateCalloutData = makeCalloutData(steps);
-
-  const newCallout = await createCallout(callout);
-  router.push({
-    path: '/admin/callouts/view/' + newCallout.slug,
-    query: { created: null },
-  });
-}
-
 onBeforeMount(async () => {
   const calloutId = route.params.id as string;
   await fetchCallout(calloutId)
-    .then((c) => Object.assign(steps, makeStepsData(c)))
-    .catch((err) => console.log('no callout found'));
+    .then((c) => {
+      console.log('found callout', c, makeStepsData(c));
+      steps.value = makeStepsData(c);
+    })
+    .catch((err) => {
+      //@ts-expect-error
+      steps.value = makeStepsData({});
+    });
 });
 </script>
